requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-agents[turn-detector]~=1.0

mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
resend


# New dependencies for enhanced features
supabase
Pillow
opencv-python
numpy
asyncio-mqtt
pydantic
pytz
timezonefinder
geopy
